DOMA-3252 delivered recipient generate functions in testScheme
