+++ apps/condo/domains/acquiring/utils/testSchema/index.js
@@ -46,9 +46,7 @@ const {
 } = require('@condo/domains/acquiring/gql')
 const { PaymentsFilterTemplate: PaymentsFilterTemplateGQL } = require('@condo/domains/acquiring/gql')
 const { Recipient: RecipientGQL } = require('@condo/domains/acquiring/gql')
-const { createValidRuBic } = require('@condo/domains/acquiring/utils/validate/bic.utils')
-const { createValidBankAccount } = require('@condo/domains/acquiring/utils/validate/bankAccount.utils')
-const { createValidTin } = require('@condo/domains/acquiring/utils/validate/tin.utils')
+const { createValidRecipient } = require('@condo/domains/acquiring/utils/testSchema/recipientGenerate')
 /* AUTOGENERATE MARKER <IMPORT> */
 
 const AcquiringIntegration = generateGQLTestUtils(AcquiringIntegrationGQL)
@@ -403,29 +401,6 @@ async function registerMultiPaymentForOneReceiptByTestClient(client, receipt, ac
     return [data.result, attrs]
 }
 
-function createValidRecipient (extra = {}) {
-    const range = (length) => ({ min: Math.pow(10,length - 1), max: Math.pow(10,length)-1 })
-
-    const bic = createValidRuBic()
-    const bankAccount = createValidBankAccount(bic)
-    const tin = createValidTin()
-
-    const validRecipient = {
-        name: faker.company.companyName(),
-        tin,
-        iec: faker.datatype.number(range(9)).toString(),
-        bic,
-        bankAccount,
-        bankName: faker.company.companyName(),
-        territoryCode: faker.datatype.number().toString(),
-        offsettingAccount: faker.finance.account(20).toString(),
-    }
-    return {
-        ...validRecipient,
-        ...extra,
-    }
-}
-
 async function createTestRecipient (client, organization, extraAttrs = {}) {
     if (!client) throw new Error('no client')
     if (!organization || !organization.id) throw new Error('no organization')
@@ -673,6 +648,5 @@ module.exports = {
     updateTestPaymentsFilterTemplate,
     registerMultiPaymentForOneReceiptByTestClient,
     Recipient, createTestRecipient, updateTestRecipient,
-    createValidRecipient,
 /* AUTOGENERATE MARKER <EXPORTS> */
 }
